Final Scrabble update
My project (with the AI commented out) ran in IntelliJ. Some technical difficulties at the last minute did not allow me to edit my comments in IntelliJ, so I had to move some things around and make comments here on GitHub. This all should run, but if it doesn't, it might be because I had to move these things around without the ability to see if it ran.